## Summary

This PR introduces a significantly improved API design that eliminates redundant `And()` method calls, providing a cleaner and more natural developer experience for traffic control configuration.

### Key Improvements

- **Eliminated Redundant And() Calls**: No more `.And()` between class configurations
- **Cleaner Method Names**: More intuitive naming (e.g., `TotalBandwidth()` vs `SetTotalBandwidth()`)
- **Natural Configuration Flow**: Configure controller → configure classes → apply
- **Enhanced Filtering**: Variadic parameters for ports, IPs, and protocols
- **Class Reuse**: Incremental configuration support for the same class
- **Better Validation**: Clear, actionable error messages

### API Comparison

**Before (v0.1.0):**
```go
controller := api.New("eth0").
    SetTotalBandwidth("100Mbps").
    CreateTrafficClass("Web Services").
        WithGuaranteedBandwidth("30Mbps").
        WithBurstableTo("60Mbps").
        WithPriority(4).
        ForPort(80, 443).
    And().  // ← Redundant\!
    CreateTrafficClass("SSH Management").
        WithGuaranteedBandwidth("5Mbps").
        WithBurstableTo("10Mbps").
        WithPriority(1).
        ForPort(22).
    And().  // ← Redundant\!
    Apply()
```

**After (Improved):**
```go
tc := api.NewImproved("eth0").
    TotalBandwidth("100Mbps")

tc.Class("Web Services").
    Guaranteed("30Mbps").
    BurstTo("60Mbps").
    Priority(4).
    Ports(80, 443)

tc.Class("SSH Management").
    Guaranteed("5Mbps").
    BurstTo("10Mbps").
    Priority(1).
    Ports(22)

tc.Apply()
```

### Enhanced Features

1. **Variadic Parameters**: `Ports(80, 443, 8080)` instead of multiple calls
2. **IP Filtering**: `SourceIPs("192.168.1.0/24", "10.0.1.100")`
3. **Protocol Filtering**: `Protocols("ssh", "http", "https")`
4. **Class Reuse**: Configure the same class multiple times incrementally
5. **Comprehensive Validation**: Clear error messages for configuration issues

### Implementation Details

- **Core API**: `api/improved_api.go` - Main implementation
- **Tests**: `test/unit/improved_api_test.go` - Comprehensive test suite (100% coverage)
- **Documentation**: `docs/improved-api-guide.md` - Complete usage guide and migration instructions
- **Examples**: `examples/improved_api_demo.go` - Working examples and comparisons

### Backward Compatibility

The existing v0.1.0 API remains fully functional. This improved API is additive and doesn't break any existing code.

### Test Coverage

All tests pass with 100% coverage:
- Basic usage scenarios
- Multiple class configurations
- Class reuse and incremental configuration
- Comprehensive validation testing
- Complex real-world scenarios
- Performance benchmarks

## Test plan

- [x] Unit tests for improved API (100% coverage)
- [x] Integration with existing codebase
- [x] Documentation and examples complete
- [x] Backward compatibility verified
- [x] Performance benchmarks included
- [x] Real-world usage scenarios tested

🤖 Generated with [Claude Code](https://claude.ai/code)